Put Animator into separate package because it needs other utilities
