+++ plugins/hu.elte.txtuml.diagnostics/src/hu/elte/txtuml/diagnostics/session/ModelMapper.java
@@ -24,7 +24,7 @@ import hu.elte.txtuml.export.uml2.mapping.ModelMapProvider;
  * Keeps track of model mappings
  * @author gerazo
  */
-class ModelMapper {
+public class ModelMapper {
 	private static final String MAPPING_FILE_EXTENSION = ModelMapProvider.MAPPING_FILE_EXTENSION;
 	private static final String MAPPING_FILE_EXTENSION_TOKEN = "." + MAPPING_FILE_EXTENSION;
 	private static final String MAPPING_DIRECTORY_PATH =
@@ -71,7 +71,7 @@ class ModelMapper {
 		mappingFileNameToMapProvider = null;
 	}
 	
-	EObject getEObjectForModelAndElement(String modelClassName, String elementClassName) {
+	public EObject getEObjectForModelAndElement(String modelClassName, String elementClassName) {
 		ModelMapProvider modelMapProvider = null;
 		Map.Entry<String, ModelMapProvider> entry = mappingFileNameToMapProvider.floorEntry(modelClassName);
 		// We do depend heavily on the fact that the mapping file name begins
