+++ plugins/hu.elte.txtuml.diagnostics/src/hu/elte/txtuml/diagnostics/session/UniqueInstance.java
@@ -13,7 +13,7 @@ import hu.elte.txtuml.diagnostics.PluginLogWrapper;
  * object in it.
  * @author gerazo
  */
-class UniqueInstance {
+public class UniqueInstance {
 	// unique ID
 	final String classInstanceID;
 	final int serviceInstanceID;
